Move boolean2 engine files up to src/; rename to boolean2_* form
