+++ src/boolean2_winding.cpp
@@ -33,12 +33,12 @@
 #include <utility>
 #include <vector>
 
-#include "../../parallel.h"
-#include "canonicalize.h"
-#include "diagnostics.h"
+#include "boolean2_diagnostics.h"
+#include "cross_section/boolean2/canonicalize.h"
+#include "cross_section/boolean2/predicates.h"
+#include "cross_section/boolean2/winding_filter.h"
 #include "manifold/optional_assert.h"
-#include "predicates.h"
-#include "winding_filter.h"
+#include "parallel.h"
 
 namespace manifold {
